Consolidate repository details, fix state listener
